Added Products create,show,edit,delete

+++ app/Http/Controllers/ProductController.php
@@ -23,7 +23,7 @@ class ProductController extends Controller
         $validated = $validator->validated();
         $query = Product::where(
             'user_id',
-            auth()->user()->id
+            Auth::user()->id
         );
         if(array_key_exists('search', $validated)){
             $search = '%' . $validated['search'] . '%';
@@ -42,9 +42,23 @@ class ProductController extends Controller
 
     public function store(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required|string|min:2|max:255',
+            'price' => 'required|decimal:0,2|min:0.01|max:999999.99'
+        ], attributes: [
+            'name' => 'nombre',
+            'price' => 'precio'
+        ]);
+        if($validator->fails()){
+            return redirect()
+                ->route('products.create')
+                ->withErrors($validator)
+                ->withInput();
+        }
+        $validated = $validator->validated();
         $product = Product::create([
-            'name' => $request->get('name'),
-            'price' => $request->get('price'),
+            'name' => $validated['name'],
+            'price' => $validated['price'],
             'user_id' => Auth::user()->id
         ]);
         return redirect()->route('products.show', $product->id);
@@ -52,6 +66,7 @@ class ProductController extends Controller
 
     public function show(Product $product)
     {
+        $this->authorize($product);
         return view('entities.products.show', [
             'product' => $product
         ]);
@@ -59,6 +74,7 @@ class ProductController extends Controller
 
     public function edit(Product $product)
     {
+        $this->authorize($product);
         return view('entities.products.edit', [
             'product' => $product
         ]);
@@ -66,16 +82,39 @@ class ProductController extends Controller
 
     public function update(Request $request, Product $product)
     {
+        $this->authorize($product);
+        $validator = Validator::make($request->all(), [
+            'name' => 'required|string|min:2|max:255',
+            'price' => 'required|decimal:0,2|min:0.01|max:999999.99'
+        ], attributes: [
+            'name' => 'nombre',
+            'price' => 'precio'
+        ]);
+        if($validator->fails()){
+            return redirect()
+                ->route('products.edit', $product->id)
+                ->withErrors($validator)
+                ->withInput();
+        }
+        $validated = $validator->validated();
         $product->update([
-            'name' => $request->get('name'),
-            'price' => $request->get('price')
+            'name' => $validated['name'],
+            'price' => $validated['price']
         ]);
         return redirect()->route('products.show', $product->id);
     }
 
     public function destroy(Product $product)
     {
+        $this->authorize($product);
         $product->delete();
         return redirect()->route('products.index');
+    }
+
+    private function authorize(Product $product): void
+    {
+        if($product->user_id !== Auth::user()->id){
+            abort(403);
+        }
     }
 }
